moved interfaces to separate files
