+++ src/GitLink/GitLink.csproj
@@ -97,6 +97,9 @@
     <Compile Include="Extensions\ProjectItemExtensions.cs" />
     <Compile Include="Exceptions\GitLinkException.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
+    <Compile Include="Interfaces\IProjectConfigurationInSolution.cs" />
+    <Compile Include="Interfaces\IProjectInSolution.cs" />
+    <Compile Include="Interfaces\ISolutionParser.cs" />
     <Compile Include="Helpers\PdbStrHelper.cs" />
     <Compile Include="Helpers\ProjectHelper.cs" />
     <Compile Include="Helpers\ResourceHelper.cs" />
