+++ src/GitLink/Interfaces/IProjectConfigurationInSolution.cs
@@ -1,0 +1,7 @@
+namespace GitLink.Interfaces
+{
+    public interface IProjectConfigurationInSolution
+    {
+        bool IncludeInBuild { get; }
+    }
+}
