+++ src/GitLink/Interfaces/IProjectInSolution.cs
@@ -1,0 +1,9 @@
+namespace GitLink.Interfaces
+{
+    public interface IProjectInSolution
+    {
+        object ProjectType { get; }
+        string RelativePath { get; }
+        object ProjectConfigurations { get; }
+    }
+}
