+++ src/GitLink/Interfaces/ISolutionParser.cs
@@ -1,0 +1,11 @@
+namespace GitLink.Interfaces
+{
+    using System.IO;
+
+    public interface ISolutionParser
+    {
+        StreamReader SolutionReader { get; set; }
+        object[] Projects { get; }
+        void ParseSolution();
+    }
+}
